# appPackage/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Azure AI Foundry Copilot API",
    "description": "API for Azure AI Foundry chat completions and conversation management, designed as a Microsoft 365 Copilot plugin.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:7181",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "oauth2Auth": [
        "api://YOUR-CLIENT-ID/access_as_user"
      ]
    }
  ],
  "paths": {
    "/api/ai-foundry/chat": {
      "post": {
        "operationId": "sendChatPrompt",
        "summary": "Send a chat prompt to Azure AI Foundry",
        "description": "Sends a user prompt to the Azure AI Foundry deployment and returns an AI-generated completion with model metadata.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI completion returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiChatResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error — prompt is empty or parameters are out of range."
          }
        }
      }
    },
    "/api/ai-foundry/unread-email-summary": {
      "post": {
        "operationId": "summarizeMyUnreadEmails",
        "summary": "Summarize unread inbox emails",
        "description": "Reads unread Microsoft 365 inbox emails through Microsoft Graph and uses Azure AI Foundry to produce a prioritized summary.",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnreadEmailSummaryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unread inbox summary returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnreadEmailSummaryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error — top/maxTokens/temperature values are out of range."
          }
        }
      }
    },
    "/api/conversations/{conversationId}": {
      "get": {
        "operationId": "getConversation",
        "summary": "Get a conversation by ID",
        "description": "Retrieves a specific conversation record by its unique identifier.",
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the conversation to retrieve.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation found and returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatConversation"
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found."
          }
        }
      }
    },
    "/api/conversations/recent": {
      "get": {
        "operationId": "listRecentConversations",
        "summary": "List recent conversations",
        "description": "Returns the most recent conversations, ordered by creation date descending.",
        "parameters": [
          {
            "name": "count",
            "in": "query",
            "required": false,
            "description": "Number of recent conversations to return. Defaults to 10.",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of recent conversations.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatConversation"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations": {
      "post": {
        "operationId": "saveConversation",
        "summary": "Save a conversation",
        "description": "Saves a conversation record containing the user prompt, AI response, and metadata.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatConversation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Conversation saved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatConversation"
                }
              }
            }
          },
          "400": {
            "description": "Validation error — ConversationId is required."
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "oauth2Auth": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://login.microsoftonline.com/common/oauth2/v2.0/authorize",
            "tokenUrl": "https://login.microsoftonline.com/common/oauth2/v2.0/token",
            "scopes": {
              "api://YOUR-CLIENT-ID/access_as_user": "Access Azure AI Foundry Copilot API as signed-in user."
            }
          }
        }
      }
    },
    "schemas": {
      "AiChatRequest": {
        "type": "object",
        "required": ["prompt"],
        "properties": {
          "prompt": {
            "type": "string",
            "description": "The user's chat prompt to send to the AI model."
          },
          "maxTokens": {
            "type": "integer",
            "description": "Maximum number of tokens in the AI response. Range: 1–4096.",
            "default": 512,
            "minimum": 1,
            "maximum": 4096
          },
          "temperature": {
            "type": "number",
            "description": "Sampling temperature for response creativity. Range: 0.0–2.0.",
            "default": 0.2,
            "minimum": 0.0,
            "maximum": 2.0
          }
        }
      },
      "AiChatResponse": {
        "type": "object",
        "required": [
          "completion",
          "model",
          "createdAtUtc",
          "sources"
        ],
        "properties": {
          "completion": {
            "type": "string",
            "description": "The AI-generated response text."
          },
          "model": {
            "type": "string",
            "description": "The model deployment name that generated the response."
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time",
            "description": "UTC timestamp when the response was generated."
          },
          "sources": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of source references used in the response."
          }
        }
      },
      "UnreadEmailSummaryRequest": {
        "type": "object",
        "properties": {
          "top": {
            "type": "integer",
            "description": "Number of unread messages to fetch from inbox. Range: 1–25.",
            "default": 5,
            "minimum": 1,
            "maximum": 25
          },
          "maxTokens": {
            "type": "integer",
            "description": "Maximum number of tokens in the AI summary. Range: 64–4096.",
            "default": 512,
            "minimum": 64,
            "maximum": 4096
          },
          "temperature": {
            "type": "number",
            "description": "Sampling temperature for summary generation. Range: 0.0–2.0.",
            "default": 0.2,
            "minimum": 0.0,
            "maximum": 2.0
          }
        }
      },
      "GraphEmailMessage": {
        "type": "object",
        "required": [
          "id",
          "subject",
          "fromName",
          "fromAddress",
          "receivedAtUtc",
          "preview"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "fromName": {
            "type": "string"
          },
          "fromAddress": {
            "type": "string"
          },
          "receivedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "preview": {
            "type": "string"
          }
        }
      },
      "UnreadEmailSummaryResponse": {
        "type": "object",
        "required": [
          "summary",
          "unreadCount",
          "emails",
          "model",
          "createdAtUtc",
          "sources"
        ],
        "properties": {
          "summary": {
            "type": "string",
            "description": "AI-generated summary and prioritization of unread inbox messages."
          },
          "unreadCount": {
            "type": "integer"
          },
          "emails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GraphEmailMessage"
            }
          },
          "model": {
            "type": "string"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "sources": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ChatConversation": {
        "type": "object",
        "required": [
          "conversationId",
          "userPrompt",
          "aiResponse",
          "createdAtUtc"
        ],
        "properties": {
          "conversationId": {
            "type": "string",
            "description": "Unique identifier for the conversation."
          },
          "userPrompt": {
            "type": "string",
            "description": "The original user prompt."
          },
          "aiResponse": {
            "type": "string",
            "description": "The AI-generated response."
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time",
            "description": "UTC timestamp when the conversation was created."
          }
        }
      }
    }
  }
}